requirements.txt
flask==3.0.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.3.2
openai==1.12.0
python-dotenv==1.0.1
plotly==5.18.0